better is_run_successfull for megarun
